Add HTTPS support
It's not meant to be for running on the public Internet. The goal is to run
autocert-stub on the public interface to get a valid certificate from Let's
Encrypt, store that in the database, and then let notes-server run reusing this
certificate on the private network.

Such configuration implies split DNS support.